chore: Update Source File Viewer
